Fetch API Task complate 100%  using useEffect,useState,UseLocation,useParam,Link and create a  function for show all data.
