+++ react/test_react_js/src/App.js
@@ -1,14 +1,15 @@
 import React from "react";
 import { BrowserRouter as Main, Route, Routes } from "react-router-dom";
-import Post from "./components/Post";
+import Allpages from "./components/Allpages";
+import Page from "./components/Page"
 // import Form from "./components/Form";
 
 function App() {
   return (
     <Main>
       <Routes>
-        <Route path="/" element={<Post />} />
-        <Route exact path="/page/:id" element={<Post  />} />
+        <Route path="/" element={<Allpages />} />
+        <Route exact path="/Page/:id" element={<Page  />} />
       </Routes>
     </Main>
   );
